\name{newLocida}
\alias{newLocida}
\title{Generate a new Locida object as R closure}
\usage{
  newLocida(X, spectLim = NULL)
}
\arguments{
  \item{X}{A standard statistical data matrix describing
  the NMR data}

  \item{spectLim}{Optional bucket limits (if not deduceable
  from X)}
}
\value{
  R closure with methods similar to Java/C++ objects
}
\description{
  Blah...
}
\details{
  Details...

  \itemize{ \item Item 1 \item Item 2 \item Item 3}
}
\author{
  Arho Virkki <arho.virkki@vtt.fi>
}
\references{
  The format of
  \href{http://www.moleculardevices.com/Documents/general-documents/mkt-appnotes/microarray-appnotes/GenePix_Pro_AppNote_Making_GAL_Files_rev_B.pdf}{
  \emph{Jack Ye Zhai: Making GenePix Array List (GAL)
  Files}}.
}

